Extract FileWatcher class
Refactoring to separate the file update logic from the top-level configuration.
