+++ backend-checks/src/main/scala/uk/gov/nationalarchives/tdr/localaws/backendchecks/checks/FileCheck.scala
@@ -1,0 +1,7 @@
+package uk.gov.nationalarchives.tdr.localaws.backendchecks.checks
+
+import java.nio.file.Path
+
+trait FileCheck {
+  def check(path: Path): Unit
+}
